Widen the character id to 64 bits in the deletion packets
Both SCCharacterDeleted and SCCharacterDeleteResponse wrote the character id as
a u32. The target client reads 8 bytes, so every packet was four bytes short and
everything after the id was shifted - the name in one, the status and both
timestamps in the other.

The name framing was already correct: a u16 byte length followed by the raw
bytes and no terminator. It now also honours the client's 128-byte bound rather
than emitting a longer name the client would reject.

Verified against the wire vectors published with the analysis - an empty name, a
short ASCII name and a multi-byte name all reproduce byte for byte.

+++ AAEmu.Game/Core/Packets/G2C/SCCharacterDeleteResponsePacket.cs
@@ -23,7 +23,8 @@ public class SCCharacterDeleteResponsePacket : GamePacket
 
     public override PacketStream Write(PacketStream stream)
     {
-            stream.Write(_characterId);
+            // Verified against the target client: the id is 8 bytes wide here, not 4.
+            stream.Write((ulong)_characterId);
             stream.Write(_status);
             stream.Write(_deleteRequestedTime);
             stream.Write(_deleteDelay);
